+++ AuraD/Src/Syntax/Keyword.cpp
@@ -5,7 +5,7 @@ Keyword::Keyword(const std::wstring & keyword): value(keyword){}
 void Keyword::Process(Cursor & cursor)
 {
 	std::wstring::const_iterator old = cursor.pos;
-	while (!cursor.AtEnd() && cursor.pos - old < value.length())
+	while (!cursor.AtEnd() && static_cast<size_t>(cursor.pos - old) < value.length())
 		++cursor.pos;
 
 	success = value == std::wstring(old, cursor.pos)/* && (cursor.AtEnd() || !isalpha(*cursor.x) && !isdigit(*cursor.x))*/;
